Preparing the extension for its own repo
