--- validmind-toc/package.json
@@ -1,147 +1,0 @@
-{
-  "name": "validmind-toc",
-  "displayName": "ValidMind ToC",
-  "version": "1.0.0",
-  "description": "Generate table of contents for VM Jupyter Notebooks based on the VSCode extension by @xelad0m.",
-  "publisher": "validbeck",
-  "author": {
-    "name": "validbeck"
-  },
-  "engines": {
-    "vscode": "^1.72.0"
-  },
-  "categories": [
-    "Notebooks",
-    "Formatters",
-    "Other"
-  ],
-  "license": "SEE LICENSE.txt",
-  "homepage": "https://github.com/validmind/documentation/",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/validmind/documentation/"
-  },
-  "icon": "validmind.png",
-  "galleryBanner": {
-    "color": "#ffffff",
-    "theme": "light"
-  },
-  "keywords": [
-    "jupyter",
-    "notebook",
-    "toc",
-    "ipynb",
-    "markdown"
-  ],
-  "activationEvents": [
-    "onCommand:validmind-toc.jupyterToc",
-    "onCommand:validmind-toc.jupyterUnToc"
-  ],
-  "main": "./out/extension.js",
-  "contributes": {
-    "commands": [
-      {
-        "command": "validmind-toc.jupyterToc",
-        "title": "+ ValidMind table of contents",
-        "category": "Jupyter"
-      },
-      {
-        "command": "validmind-toc.jupyterUnToc",
-        "title": "- ValidMind table of contents",
-        "category": "Jupyter"
-      }
-    ],
-    "menus": {
-      "editor/title": [
-        {
-          "command": "validmind-toc.jupyterToc",
-          "when": "notebookType == 'jupyter-notebook' && isWorkspaceTrusted",
-          "group": "1_run@0"
-        },
-        {
-          "command": "validmind-toc.jupyterUnToc",
-          "when": "notebookType == 'jupyter-notebook' && isWorkspaceTrusted",
-          "group": "1_run@1"
-        }
-      ],
-      "commandPalette": [
-        {
-          "command": "validmind-toc.jupyterToc",
-          "when": "notebookType == 'jupyter-notebook' && isWorkspaceTrusted"
-        },
-        {
-          "command": "validmind-toc.jupyterUnToc",
-          "when": "notebookType == 'jupyter-notebook' && isWorkspaceTrusted"
-        }
-      ]
-    },
-    "configuration": {
-      "title": "ValidMind ToC",
-      "displayName": "ValidMind ToC",
-      "properties": {
-        "validmind.tableOfContents.tableHeader": {
-          "type": "string",
-          "default": "## Contents",
-          "markdownDescription": "Defines the heading for the table of contents",
-          "scope": "resource",
-          "order": 0
-        },
-        "validmind.tableOfContents.minHeaderLevel": {
-          "type": "number",
-          "default": 2,
-          "enum": [
-            1,
-            2,
-            3,
-            4,
-            5,
-            6
-          ],
-          "markdownDescription": "Defines the **minimum** level of the notebook headers to be collected in the table of contents, defaults to 2 as the title is `h1`\t\n\nOption is overridable by ToC cell config on a per notebook basis",
-          "scope": "resource",
-          "order": 1
-        },
-        "validmind.tableOfContents.maxHeaderLevel": {
-          "type": "number",
-          "default": 4,
-          "enum": [
-            1,
-            2,
-            3,
-            4,
-            5,
-            6
-          ],
-          "markdownDescription": "Defines the **maximum** level of the notebook headers to be collected in the table of contents\t\n\nOption is overridable by ToC cell config on a per notebook basis",
-          "scope": "resource",
-          "order": 3
-        }
-      }
-    }
-  },
-  "scripts": {
-    "vscode:prepublish": "npm run compile",
-    "compile": "tsc -p ./",
-    "watch": "tsc -watch -p ./",
-    "pretest": "npm run compile && npm run lint",
-    "lint": "eslint src --ext ts",
-    "test": "node ./out/test/runTest.js"
-  },
-  "devDependencies": {
-    "@types/glob": "^8.0.0",
-    "@types/mocha": "^10.0.0",
-    "@types/node": "16.x",
-    "@types/vscode": "^1.72.0",
-    "@typescript-eslint/eslint-plugin": "^5.38.1",
-    "@typescript-eslint/parser": "^5.38.1",
-    "@vscode/test-electron": "^2.1.5",
-    "eslint": "^8.24.0",
-    "glob": "^8.0.3",
-    "mocha": "^10.0.0",
-    "typescript": "^4.8.4"
-  },
-  "dependencies": {
-    "@vscode/vsce": "^3.2.1",
-    "linqts": "^1.14.4"
-  }
-}
